+++ example/build.mjs
@@ -1,7 +1,6 @@
 import path from 'path'
 import esbuild from 'esbuild'
 import sveltePlugin from 'esbuild-svelte'
-import { typescript } from 'svelte-preprocess-esbuild'
 import preprocess from 'svelte-preprocess'
 
 /** @type {import('esbuild').Plugin} */
@@ -11,7 +10,7 @@ const dedupeSveltePlugin = {
 		build.onResolve({ filter: /^svelte(\/[\a-zA-Z0-9@-]+)?$/ }, args => ({
 			path: path.resolve(process.cwd(), 'node_modules', args.path, 'index.js')
 		}))
-	} 
+	}
 }
 
 const build = async () => {
@@ -25,14 +24,7 @@ const build = async () => {
 		minify: true,
 		plugins: [
 			sveltePlugin({
-				preprocessor: [
-					await typescript({
-						target: 'es2020'
-					}),
-					preprocess({
-						typescript: false
-					})
-				]
+				preprocessor: [preprocess()]
 			}),
 			dedupeSveltePlugin
 		]